pyproject.toml
[project]
name = "nfcpy-mynacard-py"
version = "0.0.1"
description = ""
authors = [
    {name = "okaits",email = "okaits@okaits7534.net"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "nfcpy (>=1.0.4,<2.0.0)",
]

[tool.poetry]
packages = [{include = "nfcpy_mynacard"}]


[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
pylint = "^3.3.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
